3.6: Build ostree-triggers early, don't build libacl/libattr
A recent Poky commit dropped ostree from the build root, because we
are rebuilding it here in gnome-ostree.  However, we need to have the
ostree-run-triggers binary in the build root very early.

To accomplish this, move glib very early, and build ostree with a new
special --enable-triggers-only mode.

Also, drop libattr/libacl, since we aren't doing CI on them.

Finally, drop gnome-ostree from here, because we are building it in
the Poky branch.