build: Add a "build-epoch" system that triggers rebuilds
The xorg-server ABI was just bumped, so the modules fail to load.  We
need a mechanism to simply force a rebuild of the modules, even though
they didn't change.

This is an implementation of such a mechanism.  The primary advantage
of doing it this way is that the sole input to the build system
remains git repositories - there's no need for a separate out of band
signalling mechanism with a different authentication system, etc.

The downside is that the rebuild epoch is kind of conceptually tied to
one build server.  So in theory it should probably be a git repository
tied to a builder.