Rewrite tasks system
Merge SubTask and DynTask for now - drop the dependencies, clean
things up a bit.  Move resolve, build, builddisks to be tasks.
The autobuilder now is updated to use this.

Add a new bdiff task.

There's a new "ostbuild make" builtin that runs tasks.  The benefit of
this when you do a build directly, like:

$ ostbuild make build/gnomeos-3.8

It's consistently run in a tempdir, logged etc., in the same way that
the autobuilder does it.