build: Run triggers at build time, generate initramfs
We ship a kernel, and there's really no reason *not* to build the
initramfs on the server side too.  This avoids clients doing it, which
is obviously more efficient for them.

However to do this correctly, we need to also run the triggers by
default on the server side too.  There's no major conceptual problem
with this, but it's a bit ugly because some of the file formats
include timestamps and will thus be constantly changing.

Long term fix for that: delete timestamps.