Require tools to be run from build directory, drop ~/.config/ostbuild.cfg
We want the ability to do multiple builds at once sanely; for example,
there should be a mainline autobuilder, but we should be able to
"clone" it to try a queued patch.

The "prefix" concept was kind of attempting to address this - the
vague idea was that mainline is say "gnomeos-3.8", and you'd create a
new prefix "gnomeos-3.8-bug697332" that had a patch from that bug.

But in reality things were messy because all the tools read the global
config file from ~/.config/ostbuild.cfg.

Let's approach this from another direction - a build directory is tied
to a given prefix, and if you want to do a separate build, you need to
make a separate build directory.

Then to make things efficient, later we'll add a tool to "clone" a
build directory, sharing all of the data.

For now though, it's best to just try to keep killing this idea of a
"prefix" that gets written all over the build directory.  For the
cases where we do potentially have shared data (like the repo), the
manifest already has an 'osname' we can use.