build: Write gnome-ostree/buildmaster-x86_64-{runtime,devel}
As further work on killing the "prefix" concept, the result of
a build has a new naming scheme.  Instead of

trees/gnomeos-3.8-x86_64-devel

we write:

gnome-ostree/buildmaster-x86_64-devel

The main benefit of the new scheme is that separate operating systems
can cleanly share one repo, which allows efficient branching.

The idea behind the "buildmaster" tag is that the smoketests could
create a separate gnome-ostree/smoketested-x86_64-devel tag, etc.