build: Add support for forcing a rebuild of all components
The split debuginfo work requires rebuilding everything; so
let's add support for that.

While we have the patient open, change the way builds are forced so
that we clear the component cache; this ensures that the process will
be correctly incremental.  Previously, if we detected a new epoch,
we'd save it, and if the build crashed, then the next time we started
we wouldn't rebuild anything past the crash point.