manifest: Going purely virtual
At the moment, we don't have a "maintainer" or resources for testing
gnome-ostree on physical hardware.  While I definitely understand that
the ability to do bare metal deployments of the latest GNOME could be
very useful for many types of people, we can't recommend it until a
few technical conditions are met:

1) We have testing coverage of our kernel build, particularly with
   respect to the various Linux filesystems.  Otherwise we risk
   corrupting people's drives.
2) There is someone who will keep track of core graphics stack
   components like Mesa and we have infrastructure to validate
   updates across different hardware (i.e. we have physical laptops
   with Intel chips wired into a test framework)

Besides technical, there are also political concerns:

1) Being purely virtual decreases the inherent conflict with our
   downstreams, and makes roles more clear.  Downstreams are
   responsible for "hardware enablement".
2) Some components essential for some forms of hardware (mainly
   linux-firmware) have licensing terms not acceptable to all of
   our downstreams and partners (for firmware, the FSF).

In return for solely targeting virtualization, efforts for the
near-term future (at least a year) will focus heavily on more
automated testing, which is extremely easy to do when validation is
restricted to virtualized environments.

Once we have an extensively tested system that runs well in qemu, free
of error spew on startup, memory leaks, and the like, then we can look
at what it would take to add in physical hardware targets.