task: Don't use TaskDef to run tasks in different processes
Before, we were using the TaskDef to represent both state in
the parent runner process, and things in the child process. This
got a bit confusing when understanding what code did what.

Introduce a new TaskRunner that takes a TaskDef and runs it,
splitting out the parent's state from the child's.

For now, we still instantiate TaskDef objects in both the parent
and the server, as it's the easiest way to share task metadata,
but this will soon change.