task: Split out the task metadata from the task state and code
Some tasks have metadata that's interpreter by the runner, like
the task's name, how many failed attempts it should retain, etc.
Split this out into a separate object known as a "TaskDef". The
old "TaskDef" which contains state and the code will be renamed
to a "Task", and will contain a "TaskDef" object on its prototype.