testbase: Talk to QMP monitor socket asynchronously
For future work I want to do here, I ended up with a deadlock
attempting to do synchronous communication with the guest, and gjs was
blocked on reading from the monitor socket.

Untangle this by doing read-sides of QMP commands asynchronously.  We
should also do the write side asynchronously, but that's less of an
issue.