task: Add a tasks dir that only updates on task completion
Some clients, like the IRC bot and website, may not care about the
current task that is being worked on, but instead the last one that
completed, either successfully or not. While we may want to expose
a more rich way of doing this eventually (show the last successful
or failed build for each task individiually), this is a good enough
stopgap to keep the IRC bot in semi-working order.