Add support for "child components"
A child component is a bit like a submodule - it's a directory inside a parent
module that is checked out from version control. But instead of being
configured in git, it is configured in manifest.json - the reason for
this is to allow us to rebuild the outer module whenever the inner
module changes, rather than having a fixed version of the inner module.

Child components must have unique names globally within manifest.json - this
is because they are treated like main components in certain contexts:
in particular names of child components can be passed in the
'fetchComponents' argument of 'resolve' and child components appear
independently in the bdiff output.

This will be used for the kernel build where the outer module has our
configuration and build scripts, and the inner module is the actual upstream
kernel tree.

https://bugzilla.gnome.org/show_bug.cgi?id=732395