Switch lighttpd to a tarball
This is because the fetch from git.lighttpd.org has hung several
times, stalling the whole resolve process.

Specify rm-configure because otherwise once the tarball import
process resets the timestamps, a maintainer-mode style rebuild
rules want the aclocal that the tarball was generated with,
which is newer than we have.