Add a map from revision to serial and target for each day
For perf.gnome.org, we need the ability to map back from revision
to a particular build, so we can use recorded build data to figure
out what changed between measurements.

We know within a few minutes when a build completed and measurement
started, so it's sufficient to be able to do the map at day granularity,
so add a map to each day's index.json that goes from revision to
build serial and target.

https://bugzilla.gnome.org/show_bug.cgi?id=733838